requirements.txt
python-dotenv
pandas
numpy
certifi
pymongo[srv]
scikit-learn
dill
PyYAML

# -e .